[elastic] Move the deps download into 'intialize' request handle (#79)
The original implementation relies on 'go list' to download the deps.
And 'go list' will be called by the request handler indirectly.
This approach is unpredictable and 'go list' is black box to us. So it's
better download the deps manually in 'ManageDeps', i.e. inside
'initialize' request handler.

If the client sends the initialize option `installGoDependency: true`,
download the deps in `initialize` handler.

If the client doesn't send the initialize option `installGoDependency: true`,
disable the network access by `GOPROXY=off` and find the deps from vendor
folder.